docs(mcp): Complete MCP server stub implementation and documentation
- Update AGENTS.md to reflect MCP server as stub implementation instead of disabled
- Add MCP_IMPLEMENTATION_STATUS.md with comprehensive implementation details and options
- Add MCP_SOLUTION_SUMMARY.md documenting architecture and design decisions
- Add VERIFICATION_REPORT.md with testing and validation results
- Update platform.rs, sync.rs, theme.rs, undo.rs, and virtualization.rs with minor fixes
- Add system monitoring and maintenance environment variable documentation
- Fix Cargo.toml formatting (trailing newline)
- Clarify that MCP server compiles successfully with stub methods but requires implementation for full functionality
- Document dependency analysis and three implementation options for future development

+++ src/interfaces/tui/platform.rs
@@ -47,7 +47,7 @@ pub struct TerminalCapabilities {
     pub cursor_shapes: bool,
     pub bracketed_paste: bool,
     pub focus_events: bool,
-    pub max_colors: u16,
+    pub max_colors: u32,
     pub terminal_type: String,
 }
 
@@ -294,7 +294,7 @@ impl PlatformManager {
     }
 
     /// Get maximum number of colors for color support level
-    fn get_max_colors(color_support: &ColorSupport) -> u16 {
+    fn get_max_colors(color_support: &ColorSupport) -> u32 {
         match color_support {
             ColorSupport::None => 0,
             ColorSupport::Basic16 => 16,
